requirements.txt
Flask==3.1.1
flask-cors==6.0.1
pymongo==4.13.2
python-dateutil==2.9.0.post0
